### 2020-05-23 #2242 @arfon
:wave: @ivalaginja - thanks for your submission to JOSS. Does this paper accompany a paper in a AAS journal?
