### 2020-05-23 #2242 @ivalaginja
@arfon thanks! No, this paper does not accompany a paper in any AAS journal. We tried to reduce the repository size by cleaning out the history before submission, but with limited success. If there is anything that can be done from our side concerning that issue, please let me know.
