### 2020-05-27 #2242 @xuanxu
:wave: @hayesla & @nespinoza - would you be willing to review this submission for JOSS?
I think your knowledge of Python and astrophysics/exoplanets makes both of you a great fit for this submission.
