### 2020-06-01 #2281 @whedon
[REVIEW]: ExoTiC-ISM: A Python package for marginalised exoplanet transit parameters across a grid of systematic instrument models
**Submitting author:** @ivalaginja (<a href="http://orcid.org/0000-0003-1783-5023">Iva Laginja</a>)
**Repository:** <a href="https://github.com/hrwakeford/ExoTiC-ISM" target ="_blank">https://github.com/hrwakeford/ExoTiC-ISM</a>
**Version:** v2.0.0
