### 2020-06-01 #2281 @xuanxu
@ivalaginja, @hayesla, @mattpitkin : this is the review thread for the paper. All of our communications will happen here from now on.

We aim for reviews to be completed within about 2-4 weeks, but that's not a hard deadline. Please let me know if any of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
