### 2020-06-16 #2281 @mattpitkin
@ivalaginja The installation instructions work well. I'd just recommend explicitly adding how to activate the conda environment before running setup.py:
```
conda activate exoticism
python setup.py develop
```
