### 2020-06-16 #2281 @mattpitkin
This isn't a requirement, but are there plans to have continuous integration (e.g., through [Travis](https://travis-ci.org/)) that can run your test suite? Do you know how much of the main functionality of the code the current test suite covers?
