### 2020-06-16 #2281 @mattpitkin
For some references they don't seem to render the journal. I think this is due to the `paper.bib` file containing the AAS macros (e.g., `\mnras`, `\aap`) for several papers - these macros should be replaced with the full or abbreviated journal name.

There are also a couple that need the DOI added:
 * [Claret, A](https://ui.adsabs.harvard.edu/abs/2000A%26A...363.1081C/abstract)
 * [McKinnet, W](https://conference.scipy.org/proceedings/scipy2010/mckinney.html) the DOI is 10.25080/Majora-92bf1922-00a
