### 2020-06-16 #2281 @mattpitkin
The functionality of the code, and required inputs in the configuration file, are well documented in the tutorial notebook, so I have ticked the **Functionality documentation** check box. There is no specific documentation for the API, so in the future I would recommend there is an autogenerated doc page that does contain the API, but this is not required for this review.
