### 2020-06-16 #2281 @mattpitkin
Once the references are fixed I'm happy to sign-off the review.

Just to summarise and slightly add to what I've written above, I've a couple of general recommendations but these are **not** requirements for the review:

 * add some continuous integration for the tests (maybe run the described example as a test and compare the output to the *expected* output)
 * add some autogenerated documentation that includes documentation for the API

One very minor typo fix for the paper: in the first sentence change "has" to "have".

I should just note that I ran the example on a remote machine (my laptop's running low on space, so I had to ssh into my work desktop to clone the repo there!). When I tried running the example it performed the first stage of fitting, but hung when it got to the second. I got things to work by changing to `plotting = False` in the config file, after which it ran through everything ok.
