### 2020-06-19 hrwakeford/ExoTiC-ISM#96 @ivalaginja
Increase CI test coverage
We have set up CI with GitHub Actions which runs on every open PR, however our code coverage is extremely low and should be increased in the future to ensure robustness with new code and updates.
