### 2020-06-19 hrwakeford/ExoTiC-ISM#97 @ivalaginja
Run hangs when plotting=True on remote machine
> I should just note that I ran the example on a remote machine (my laptop's running low on space, so I had to ssh into my work desktop to clone the repo there!). When I tried running the example it performed the first stage of fitting, but hung when it got to the second. I got things to work by changing to `plotting = False` in the config file, after which it ran through everything ok.
>
> _Originally posted by @mattpitkin in https://github.com/openjournals/joss-reviews/issues/2281#issuecomment-644721513_

The minimum we should do is to add a note about this in the docs, while maybe also setting the default to `plotting = False`?
