### 2020-06-19 #2281 @ivalaginja
Thanks for all your comments @mattpitkin! To address your questions:
1. I have updated the Quickstart section to include the activation of the environment, thanks for that!
2. We currently use GitHub Actions to run our test suite in every open PR, alternatively this can be run locally with `pytest`. Our current test coverage is very thin and probably in the lower single digits percent-wise. However, we aim at improving on that in the future which is why we already set up CI, so that all that's left is actually to write more senseful tests. We have an open issue for this [here](https://github.com/hrwakeford/ExoTiC-ISM/issues/96).
3. We have an issue open for auto-generated documentation [here](https://github.com/hrwakeford/ExoTiC-ISM/issues/29), we just decided to punt it to later since we ran into some problems in the first trial to set that up.
4. I changed the word in the first sentence of the paper.
5. I took out all the AAS macros in the `paper.bib`
6. I noted the problem you ran into on the remote machine and created an issue for it [here](https://github.com/hrwakeford/ExoTiC-ISM/issues/97).
7. I added the DOI for the `pandas` paper but can partout not find the DOI for the Claret 2000 paper. I have reached out to an author who recently cited it to ask whether they know the DOI, but I am not sure this will help. I will ask @hrwakeford (co-author) if she knows more. What would be our alternative if we can't find the DOI?
