### 2020-06-19 #2281 @ivalaginja
I just saw that [Doe et al.](https://ui.adsabs.harvard.edu/abs/2007ASPC..376..543D/abstract) also doesn't have a DOI, facing the same problem here.
