### 2020-06-19 #2281 @mattpitkin
Thanks for your detailed responses and for opening the issue.

I think you're right that there just aren't DOIs for Claret or Doe et al, so the references are fine without them. I'll tick of that last box for my review.
