### 2020-06-22 #2281 @hrwakeford
The reference to Claret 2000 has two sources an A&A paper and a catalog in VizieR
Paper - https://ui.adsabs.harvard.edu/abs/2000A%26A...363.1081C/abstract
VizieR - http://vizier.u-strasbg.fr/viz-bin/VizieR?-source=J/A+A/363/1081

The Sherpa package lists the following link as having the correct DOI for citation but I struggled to navigate to where that might be but this is where to start
https://zenodo.org/record/3631574#.XvC_bi-ZNhE
