### 2020-06-22 #2281 @ivalaginja
> https://zenodo.org/record/3631574#.XvC_bi-ZNhE

You're right about Sherpa, however this is only for the package itself, for which we already added the DOI; what's missing is the DOI for the paper they want us to cite.

As for Claret, it really seems like there's no DOI to be found :P
