### 2020-06-23 #2281 @hayesla
I think the pandas 2020 reference should be

```
@software{jeff_reback_2020_3644238,
  author       = {Jeff Reback and
                  Wes McKinney and
                  jbrockmendel and
                  Joris Van den Bossche and
                  Tom Augspurger and
                  Phillip Cloud and
                  gfyoung and
                  Sinhrks and
                  Adam Klein and
                  Matthew Roeschke and
                  Jeff Tratner and
                  Chang She and
                  Simon Hawkins and
                  William Ayd and
                  Terji Petersen and
                  Jeremy Schendel and
                  Andy Hayden and
                  Marc Garcia and
                  MomIsBestFriend and
                  Vytautas Jancauskas and
                  Pietro Battiston and
                  Skipper Seabold and
                  chris-b1 and
                  h-vetinari and
                  Stephan Hoyer and
                  Wouter Overmeire and
                  alimcmaster1 and
                  Mortada Mehyar and
                  Kaiqi Dong and
                  Christopher Whelan},
  title        = {pandas-dev/pandas: Pandas 1.0.1},
  month        = feb,
  year         = 2020,
  publisher    = {Zenodo},
  version      = {v1.0.1},
  doi          = {10.5281/zenodo.3644238},
  url          = {https://doi.org/10.5281/zenodo.3644238}
}
```
