### 2020-06-24 #2281 @hayesla
Overall looks good to me. The Quickstart guide worked well, and I was able to go through the example notebook easily after this and play around with the codebase.

Once the reference (above) is fixed I have nothing blocking and just some suggested comments

- I agree with @mattpitkin, a specific description of the API, or even just an overview in the README would be good to make it clear what is available. And of course something like an automated documentation of the API, maybe https://readthedocs.org/ is something to consider for the future.

- For the **Community Guidelines** I would suggest adding a few sentences to the README explicitly mention where users can report issues or problems seek support, maybe within another heading?

- The tutorial notebook is great and is well documented to describe the functionality of the code and apply it to some nice examples. I would certainly suggest moving this to some documentation somewhere else in the future also.

- Future continuous integration functionality would be great to see.
