### 2020-06-24 #2281 @ivalaginja
@hayesla thank you for your comments! I fully agree with you on all you said, and we will work in the near future to get these things set up; issues for this already exist in the repository.

As for the reference, thanks for pointing it out. I will for sure update the author list, however, the DOI and the version wouldn't work for our paper since we used pandas version v0.24.2, for which there is no explicit Zenodo entry. My idea was to stick in the concept DOI instead, which will always point to the most recent release since there is no DOI for the particular version we use. And I could adjust the version number in the Bibtex manually.

Would that be an acceptable solution?
