### 2020-06-24 #2281 @hayesla
I'm not too sure what you mean - as the DOI that is listed in paper for (`team, T. pandas development. (2020). Pandas-dev/pandas: Pandas 1.0.1. Zenodo. doi:10.
5281/zenodo.3644238` ) to here https://zenodo.org/record/3644238#.XvNnWGpKhTY which in the bottom left hand corner has a bibtex entry

Maybe I am misunderstanding?
