### 2020-06-24 #2281 @ivalaginja
If you look on the right side of that very page, in the box labelled "Versions", you will see a list of all the published versions of pandas. Each version has a unique DOI that should be cited depending on the version you use. We are using a version that was not separately published (0.24.2), so it has no unique DOI. At the bottom of this box the "concept DOI" is shown, with additional information on how that is handled. Since there is no DOI for the version we use, I thought it would be better to cite the concept DOI rather than wrongly point to a version we are not actually using?
