### 2020-06-24 #2281 @xuanxu
The “concept” DOI is 10.5281/zenodo.3509134
I think that’s the one you want to use here. In the `environment.yml` file there is no version specified for Pandas, so it makes sense to use the DOI for “all versions”.
