### 2020-06-26 #2281 @xuanxu
@ivalaginja I have issued a couple pull requests in the ExoTiC-ISM repo with minor changes. Please take a look at them and consider merging them if you think they are OK.
