### 2020-06-26 #2281 @ivalaginja
@xuanxu I had a look at them, double-checked and merged.
