### 2020-06-26 #2281 @xuanxu
OK @ivalaginja, everything looks good now, here are the next steps:

- Please release a new [tagged version](https://github.com/hrwakeford/ExoTiC-ISM/releases) from the current main branch so it includes all the changes made during the review process
- Then archive that latest release in Zenodo (or a similar service).
- Check the Zenodo deposit has the correct metadata: title and author names should match the paper; you may also add authors ORCID.

Once you do that please report here the version number and archive DOI
