### 2020-06-30 #2281 @danielskatz
👋 @ivalaginja - here are some suggested changes for the paper:

- "with the goal to identify" <- "with the goal of identifying"
- "For these giant, close-in exoplanets the most" <- "For these giant, close-in exoplanets, the most"
- "to be well mixed" <- "to be well-mixed"
- "systematic models which yields" <- "systematic models, which yields"
- "There are a number of Python solutions to create and fit transiting planet light curves, however ExoTiC-ISM" <- "Where there are a number of Python solutions to create and fit transiting planet light curves, ExoTiC-ISM"
- "(Wakeford et al., 2016), however, can be adapted to other instruments." <- "(Wakeford et al., 2016). However, this can be adapted to other instruments."
- "Freeman et al., 2001) as well as" <- "Freeman et al., 2001), as well as"
- "The development in Python" <- "Development in Python"

In addition, please fix the cases in the references, for example
"Astronomical data analysis software and systems xvi, Astronomical society of the pacific conference series"
You will need to do this by protecting (with {}s) cases in the .bib file.  A bunch of journals also need to be fixed, as we don't know or expand standard astro abbreviations, such as aap, apj, apjl, mnras, etc.

once you've done this, please issue ```@whedon generate pdf``` (and iterate if needed), then let me know it's ready to go.
