### 2020-07-03 #2281 @danielskatz
@ivalaginja - this is just a reminder that we are waiting on the changes to the paper and bib as requested in the issue above, then your paper can be published
